Relationship defined

+++ database/migrations/2019_05_03_141303_create_replies_table.php
@@ -16,8 +16,8 @@ class CreateRepliesTable extends Migration
         Schema::create('replies', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->text('body');
-            $table->integer('question_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('question_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
             $table->timestamps();
         });
